Move component entry points into their respective folders
