+++ libs/@hashintel/ds-components/src/components/Avatar/avatar.tsx
