+++ libs/@hashintel/ds-components/src/components/Badge/badge.tsx
