+++ libs/@hashintel/ds-components/src/components/Button/button.stories.tsx
@@ -1,6 +1,6 @@
 import type { Story, StoryDefault } from "@ladle/react";
 
-import { Button, type ButtonProps } from "../button";
+import { Button, type ButtonProps } from "./button";
 
 export default {
   title: "Components/Button",
