+++ libs/@hashintel/ds-components/src/components/SegmentedControl/segmented-control.tsx
