+++ libs/@hashintel/ds-components/src/components/Slider/slider.tsx
